+++ bk-api/app.go
@@ -1,7 +1,8 @@
 package main
 
 import (
-	_ "bk-api/controller"
+	_ "bk-api/controller/api"
+	_ "bk-api/controller/app"
 	_ "bk-api/log"
 	"net/http"
 	"github.com/mageddo/go-logging"
@@ -9,10 +10,9 @@ import (
 
 func main(){
 
-	logger := logging.NewLog(logging.NewContext())
-	logger.Debugf("status=starting...")
+	logging.Debugf("status=starting...")
 	if err := http.ListenAndServe(":3131", nil); err != nil {
-		logger.Errorf("error=%v", err)
+		logging.Errorf("error=%v", err)
 	}
 }