+++ bk-api/dao/utils_dao.go
@@ -1,0 +1,9 @@
+package dao
+
+type UtilsDAO interface {
+	Exec(sql string)
+}
+
+func NewUtilsDAO() *UtilsDAOSQLite {
+	return &UtilsDAOSQLite{}
+}