+++ bk-api/entity/setting.go
@@ -1,0 +1,9 @@
+package entity
+
+import "time"
+
+type SettingEntity struct {
+	Key string `json:"key"`
+	Value string `json:"value"`
+	UpdateDate *time.Time `json:"updateDate"`
+}